+++ packages/registry/src/index.ts
@@ -37,6 +37,7 @@ export * from '../registry/default/loading-page';
 export * from '../registry/default/meter-row';
 export * from '../registry/default/modal';
 export * from '../registry/default/multi-stat-card';
+export * from '../registry/default/number-input';
 export * from '../registry/default/page-header';
 export * from '../registry/default/pagination';
 export * from '../registry/default/popover';